+++ backend/canisters/local_user_index/impl/src/jobs/mod.rs
@@ -9,6 +9,7 @@ pub mod topup_canister_pool;
 pub mod topup_canisters;
 pub mod upgrade_communities;
 pub mod upgrade_groups;
+pub mod upgrade_multi_users;
 pub mod upgrade_users;
 
 pub(crate) fn start(state: &RuntimeState) {
@@ -19,6 +20,7 @@ pub(crate) fn start(state: &RuntimeState) {
     topup_canisters::start_job();
     upgrade_communities::start_job_if_required(state);
     upgrade_groups::start_job_if_required(state);
+    upgrade_multi_users::start_job_if_required(state);
     upgrade_users::start_job_if_required(state);
 }
 
@@ -26,7 +28,8 @@ fn clear_chunk_store_if_no_pending_upgrades() {
     if let Some(canister_id) = read_state(|state| {
         let should_clear_chunk_store = state.data.users_requiring_upgrade.is_empty()
             && state.data.groups_requiring_upgrade.is_empty()
-            && state.data.communities_requiring_upgrade.is_empty();
+            && state.data.communities_requiring_upgrade.is_empty()
+            && state.data.multi_users_requiring_upgrade.is_empty();
 
         if should_clear_chunk_store { Some(state.env.canister_id()) } else { None }
     }) {