+++ backend/canisters/local_user_index/impl/src/model/mod.rs
@@ -4,6 +4,7 @@ pub mod global_user_map;
 pub mod group_event_batch;
 pub mod local_community_map;
 pub mod local_group_map;
+pub mod local_multi_user_map;
 pub mod local_user_map;
 pub mod media_scan_job_log;
 pub mod moderation_queue;