+++ backend/tools/canister_wasm_chunks_uploader/src/lib.rs
@@ -76,6 +76,25 @@ pub async fn upload_user_canister_wasm(
     .unwrap();
 }
 
+pub async fn upload_multi_user_canister_wasm(
+    identity: Box<dyn Identity>,
+    url: String,
+    user_index_canister_id: CanisterId,
+    version: BuildVersion,
+) {
+    let agent = build_ic_agent(url, identity).await;
+    let canister_wasm = get_canister_wasm(CanisterName::MultiUser, version);
+
+    user_index_canister_client::upload_wasm_in_chunks(
+        &agent,
+        &user_index_canister_id,
+        &canister_wasm.module,
+        user_index_canister::ChildCanisterType::MultiUser,
+    )
+    .await
+    .unwrap();
+}
+
 pub async fn upload_group_canister_wasm(
     identity: Box<dyn Identity>,
     url: String,